fix and sort gradingperiods
